lflows: Create a separate node for lflow syncing.
The goal of this is to allow for the syncing node to be able to take
logical flow tables from multiple sources and sync all of them with the
southbound database.

This essentially moves the lflow_table_sync_to_sb() function into a
separate incremental node from en-lflow. This way, en-lflow generates a
logical flow table, and en-lflow-sync syncs it with the southbound
database.

This commit slightly changes the logic of lflow_table_sync_to_sb() so
that it does not directly delete southbound logical flows. Instead, this
happens at the end of the syncing process. This gives the basis to allow
for multiple logical flow tables to be synced without prematurely
deleting southbound logical flows.

Signed-off-by: Mark Michelson <mmichels@redhat.com>