lflow_ref: Allow multiple datapaths per lflow_ref_node.
Prior to this commit, an lflow_ref_node would be created the first time
that an lflow was added to an lflow_ref. If this lflow was added for a
single datapath, then the lflow_ref_node would reference that single
datapath. If the lflow was added for a datapath_group, then the
lflow_ref_node would reference all datapaths in the group.

If the same lflow were added to the same lflow_ref a second time, but
specified a different datapath than the first time the lflow was added,
then the lflow_ref_node remained unchanged, even though the lflow now
applied to one or more new datapaths.

In most cases, this is fine, because lflows are only added once per
lflow_ref. The IGMP lflow_ref, on the other hand, works differently.
There is only a single lflow_ref for all of IGMP. Therefore, it is
possible (probable, even) for the same lflow to be added multiple times
to the same lflow_ref, each time for a different datapath.

Most lflow input node handlers start by calling lflow_ref_unlink_lflows().
This removes all of the datapath references from the lflow_ref and sets all
lflow_ref_nodes within the lflow_ref to be unlinked. Then the handler re-builds
the relevant lflows, causing any corresponding lflow_ref_nodes to be re-linked,
based on the datapath the lflow was added for. Finally, the handler calls
lflow_ref_sync_lflows(). This causes all lflows in the lflow_ref to be
synced with the southbound Logical_Flow table. Any lflow_ref_nodes that were
unlinked and not re-linked will either have their lflow's corresponding
Logical_Flow datapath-related columns updated, or they will have their
record deleted entirely if the lflow no longer has any datapath references.

This particular workflow breaks in the case that the lflow was added
multiple times to the lflow_ref, each with a different datapath. The
lflow_ref_node is only aware of the first datapath for which the lflow
was added. So when the lflow_ref_node is unlinked, only the reference to
that particular datapath is removed from the lflow. Therefore, the lflow
may still have references to datapaths in its dpg_bitmap field that
should have been removed when the lflow_ref_node was unlinked. After
rebuilding the flows and re-syncing, those leftover datapaths will still
be in the lflow, resulting in the southbound Logical_Flow referencing
datapaths for which the flow should have been removed. This translates
to the datapath in question having the lflow installed when it should
have been removed.

The IGMP lflow handler gets around this by calling
lflow_ref_resync_flows() instead of lflow_ref_unlink_lflows() at the
beginning. This causes the lflow_ref_nodes to be unlinked and synced
immediately before attempting to rebuild the lflows. Since no
lflow_ref_nodes had a chance to be re-linked, it results in all of the
lflows in the IGMP lflow_ref being deleted (both in memory and in the
Southbound Logical_Flow table), then the lflows are rebuilt from scratch.
Since the lflows are re-created entirely, there is no possibility of stray
datapaths being referenced.

The next change in this series is going to separate lflow_ref
unlinking/rebuilding into one engine node and lflow_ref syncing into a
separate one. Because of that, the en-lflow IGMP input handler can no longer
call lflow_ref_resync_flows(). It needs to use lflow_ref_unlink_lflows() like
the other lflow input handlers do. But as stated in an earlier paragraph, this
causes stray datapaths to be referenced by the southbound Logical_Flow table.

This change modifies how lflow_ref_nodes handle their datapath
references. lflow_ref_nodes can now have their datapath references updated
when an lflow is added to an lflow_ref multiple times. To accomplish this, the
lflow_ref_node dgrp_bitmap and dp_index fields have been combined into a
bitmap field, dp_bitmap. Whether a single datapath or a datapath group is
referenced by the added lflow, we handle it the same in the lflow_ref_node, by
using the dp_bitmap.

Signed-off-by: Mark Michelson <mmichels@redhat.com>