lflows: Separate lflow_ref building and syncing.
This commit gives responsibility for syncing lflow_refs to the
en-lflow-sync engine node. Currently it only handles lflow_refs from the
en-lflow node, but it could potentially handle lflow_refs from any input
node.

To accomplish this, all en-lflow input handlers mostly work the same.
The difference is that instead of calling lflow_ref_sync_lflows(), they
push their lflow_ref to a vector. The en-lflow-sync node then iterates
over this vector and calls lflow_ref_sync_lflows() on all of the
lflow_refs.

In order to ensure that all unlinked lflow_ref_nodes have their
corresponding lflows removed from the lflow_table, we no longer
short-circuit when a call to lflow_ref_sync_lflows() fails. We allow all
of the calls to complete, and afterwards if there were any failures, we
fall back to a recompute. This way, when the recompute happens, there is
no chance of attempting to sync unlinked lflows with the southbound
database.

This commit also adds an awkward call to clear datapath groups at the
beginning of lflow_table_sync_to_sb(). This is necessary because when we
fall back to a recompute from the incremental case, the dp_groups need
to be cleared out. A more natural place for this would be in the
en-lflow-sync run() function, but the internals of the lflow_table are
not accessible from the en-lflow.c file. The next commit will separate
the dp_groups from the lflow_table.

Signed-off-by: Mark Michelson <mmichels@redhat.com>