lflows: Move dp_groups from lflow_data to lflow_sync_data.
dp_groups are handled entirely within the en-lflow-sync node, so there's
no need for the en-lflow node to own them. This commit moves the
dp_groups to the en-lflow-sync node, using an array to group the
dp_groups by datpath type. We also store the relevant ovn_datapaths in
the same array index. This allows for us to more easily pass the correct
dp_groups and datapaths around.

Signed-off-by: Mark Michelson <mmichels@redhat.com>