datapaths: Add a hashvec of datapath bindings.
The ovn_datapaths structure maintains an array of datapaths, allowing
for bitmaps to be used to indicate the relevant datapaths associated
with an entity, such as a logical flow.

Logical flow syncing is attempting to rid itself of dependencies on
ovn_datapaths, but the logical flow creation hinges on the datapath
indices.

This commit addresses the issue by introducing a type-agnostic
ovn_datapath_binding_hashvec structure that can be used to store
southbound Datapath_Bindings in both a hashmap and an array. This allows
for us to correlate a southbound Datapath_Binding with a specific index
into its corresponding vector. The ovn_datapaths array is created based
on this vector, meaning that an ovn_datapath's index corresponds to the
same southbound Datapath_Binding's ovn_datapath_binding index.

In order to maintain consistency of indices, this commit also changes
synced datapath bindings to be a vector instead of a list, and sorts the
vector with each run. This ensures that synced datapaths will contain
the same indices as long as new datapaths are not added or removed from
the database. If datapaths are added or removed, then northd has to
recompute, meaning that the ovn_datapaths will recreate their arrays and
indices.

Signed-off-by: Mark Michelson <mmichels@redhat.com>