lflow: Remove northd dependency from en-lflow-sync.
The previous commit established that ovn_datapath array indices
correspond with synced logical datapath array indices. Because of that,
we can use the synced datapath indices when syncing logical flows now.

This removal of the northd dependency frees up new nodes to be able to
create logical flow tables using the synced datapath indices instead of
ovn_datapath indices, and the syncing will work no matter which set of
indices are used.

Currently, the macros to add logical flows to the logical flow table
still require an ovn_datapath but we'll change that in the next commit.

This change expanded out beyond the lflow code since en-sync-sb uses
ovn_dp_groups. Luckily, the only thing the en-sync-sb needed from
northd's ovn_datapaths was the datapath indices. This meant the
conversion to using the synced datapath indices very easy.

Signed-off-by: Mark Michelson <mmichels@redhat.com>