lflows: Remove ovn_datapath parameter when adding lflows.
This commit makes lflow_table_add_lflow no longer require an
ovn_datapath to be passed in. Instead, a datapath index value can be
used for the same purpose.

This patch compromises by only updating the core function, not the
macros that northd tends to use to add logical flows. Those macros
could be updated, but it would make the diff much larger, and it make
backports more difficult in northd.c.

Signed-off-by: Mark Michelson <mmichels@redhat.com>