northd: Fix TCP connection in stateless nat.
FIN or FIN-ACK packets from client was dropping because the client
side was using conntrack. Connection is in SYN_SENT state because the
response from backends bypass the conntrack, when client sends a FIN
or FIN-ACK, the conntrack is invalid and packet is dropped.
To fix it, remove the client side from conntrack, calculating the hash
from packet to choice the backend.
REG_IDX_LB_STATELESS is used to store the index from backend.

Signed-off-by: Lucas Vargas Dias <lucas.vdias@luizalabs.com>