northd, ic: Add hub-spoke options to adv DR routes learned in ovn-ic.
Consider the scenario where there are 2 AZs using ovn-ic, each AZ has
one LR and its LS. The LRs are interconnected by a transit switch.
On each AZ, we can configure the dynamic routing and exchange routes
via BGP with external BGP speakers.

AZ1
LS1 - LR1 - BGP Speaker1 - Local subnet1
       \
   -----------------------
    transit switch
   -----------------------
AZ2    /
LS2 - LR2 - BGP Speaker2 - Local subnet2

The LR2 will learn the subnet1 prefix via ovn-ic but this prefix will
not be redistributed to BGP Speaker2 and it also happens with LR1.
This scenario uses the network's hub-and-spoke terminology where we can
enable the hub-spoke on the LR for such redistribution.
subnet1 and subnet2 are configured as ic-source-dynamic=true in ovn-ic.
Also, consider the same scenario, but LR1 and LR2 learn the same subnet
for redundancy, hub-spoke option is disabled to not adv the subnet learned
from BGP and advertised in ovn-ic in BGP speakers.

Signed-off-by: Lucas Vargas Dias <lucas.vdias@luizalabs.com>