controller: Avoid adv the same ip prefix and priority.
If the LRPs are scheduled in the same chassis for some reason,
ovn controller will try to advertise many times the same prefix with the same priority.
This generates a high CPU load in ovn-controller and generate a lot of messages
like the following:
"Add route table_id=1337 dst=42.10.10.14 plen=32 nexthop=(blackhole) failed: File exists"
To fix it, just check if prefix was already advertised with the same priority.

Fixes: ccb0b6b9109c ("controller: Introduce route node.")
Signed-off-by: Lucas Vargas Dias <lucas.vdias@luizalabs.com>